+++ console-app/src/main/java/com/infoshareacademy/Route.java
@@ -1,6 +1,11 @@
 package com.infoshareacademy;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
 import javax.validation.constraints.NotNull;
 import java.util.Date;
 
@@ -11,7 +16,7 @@ public class Route {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
 
     @Column(name = "start_at")
     @NotNull
@@ -55,7 +60,7 @@ public class Route {
     }
 
 
-    public Route(Integer id, Date date, String startCity, String startStreet, String endCity,
+    public Route(Long id, Date date, String startCity, String startStreet, String endCity,
                  String endStreet, String pickUpCity, String pickUpStreet, String startTime, String endTime, String pickUpTime) {
         this.id = id;
         this.date = date;
@@ -70,7 +75,7 @@ public class Route {
         this.pickUpTime = pickUpTime;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
@@ -112,7 +117,7 @@ public class Route {
         return pickUpTime;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
